+++ src/components/Error.js
@@ -1,7 +1,7 @@
 function Error() {
   return (
     <p className="error">
-      <span>💥</span> There was an error fecthing questions.
+      <span>💥</span> There was an error fetching questions.
     </p>
   );
 }
