+++ src/components/Progress.jsx
@@ -1,14 +1,24 @@
-import React from 'react'
+import React from "react";
 
-const Progress = ({index, numQuestions,points, totalPossiblePoints, answer}) => {
+const Progress = ({
+  index,
+  numQuestions,
+  points,
+  totalPossiblePoints,
+  answer,
+}) => {
   return (
-    <header className='progress'>
-      <progress max={numQuestions} value={index + Number(answer !== null)}/>
+    <header className="progress">
+      <progress max={numQuestions} value={index + Number(answer !== null)} />
 
-      <p>Question <strong>{index + 1}</strong> / {numQuestions}</p>
-      <p><strong>{points}</strong> / {totalPossiblePoints}</p>
+      <p>
+        Question <strong>{index + 1}</strong> / {numQuestions}
+      </p>
+      <p>
+        <strong>{points}</strong> / {totalPossiblePoints}
+      </p>
     </header>
-  )
-}
+  );
+};
 
-export default Progress
+export default Progress;
