+++ src/components/Question.jsx
@@ -1,15 +1,12 @@
 import React from "react";
 import Options from "./Options";
 
-const Question = ({questions, dispatch,answer}) => {
- 
- 
-
+const Question = ({ questions, dispatch, answer }) => {
   return (
     <div>
       <h4>{questions.question}</h4>
 
-      <Options questions={questions} dispatch={dispatch} answer={answer}/>
+      <Options questions={questions} dispatch={dispatch} answer={answer} />
     </div>
   );
 };
