+++ src/components/Timer.jsx
@@ -1,21 +1,22 @@
-import React, { useEffect } from 'react'
+import React, { useEffect } from "react";
 
-const Timer = ({dispatch, secondsRemaining}) => {
-
-  const min = Math.floor(secondsRemaining / 60)
-  const sec = secondsRemaining % 60
+const Timer = ({ dispatch, secondsRemaining }) => {
+  const min = Math.floor(secondsRemaining / 60);
+  const sec = secondsRemaining % 60;
 
   useEffect(() => {
-  const id =  setInterval(() => {
-      dispatch( {type: "timer"})
-    }, 1000)
-    return () => clearInterval(id)
-  },[dispatch, secondsRemaining])
+    const id = setInterval(() => {
+      dispatch({ type: "timer" });
+    }, 1000);
+    return () => clearInterval(id);
+  }, [dispatch, secondsRemaining]);
   return (
-    <div className='timer'>
-    {min < 10 && 0}{min} : {sec < 10 && 0}{sec}
+    <div className="timer">
+      {min < 10 && 0}
+      {min} : {sec < 10 && 0}
+      {sec}
     </div>
-  )
-}
+  );
+};
 
-export default Timer
+export default Timer;
